+++ src/plugins/dev-mode/unallowed-properties.ts
@@ -1,9 +1,7 @@
 import type { RxCollectionCreator, RxDatabaseCreator } from '../../types';
-import { newRxError } from '../../rx-error';
+import { newRxError, newRxTypeError } from '../../rx-error';
 import { rxDatabaseProperties } from './entity-properties';
 import { isFolderPath } from '../../util';
-import { validateDatabaseName } from './check-names';
-
 
 /**
  * if the name of a collection
@@ -18,6 +16,7 @@ export function ensureCollectionNameValid(
             name: args.name
         });
     }
+    validateDatabaseName(args.name);
 }
 
 export function ensureDatabaseNameIsValid(args: RxDatabaseCreator<any, any>) {
@@ -36,5 +35,44 @@ export function ensureDatabaseNameIsValid(args: RxDatabaseCreator<any, any>) {
             });
         }
     }
+}
 
+
+
+const validCouchDBStringRegexStr = '^[a-z][_$a-z0-9\\-]*$';
+const validCouchDBStringRegex = new RegExp(validCouchDBStringRegexStr);
+
+/**
+ * Validates that a given string is ok to be used with couchdb-collection-names.
+ * We only allow these strings as database- or collection names because it ensures
+ * that you later do not get in troubble when you want to use the database together witch couchdb.
+ *
+ * @link https://docs.couchdb.org/en/stable/api/database/common.html
+ * @link https://neighbourhood.ie/blog/2020/10/13/everything-you-need-to-know-about-couchdb-database-names/
+ * @throws  {RxError}
+ */
+export function validateDatabaseName(name: string): true {
+    if (
+        typeof name !== 'string' ||
+        name.length === 0
+    ) {
+        throw newRxTypeError('UT1', {
+            name
+        });
+    }
+
+
+    // do not check, if foldername is given
+    if (isFolderPath(name)) {
+        return true;
+    }
+
+    if (!name.match(validCouchDBStringRegex)) {
+        throw newRxError('UT2', {
+            regex: validCouchDBStringRegexStr,
+            givenName: name,
+        });
+    }
+
+    return true;
 }